requirements.txt
python-telegram-bot==20.0
requests==2.26.0
python-dotenv==0.19.2
pymongo[srv]==4.6.1